requirements.txt
matplotlib
pandas
python-dotenv
selenium
